% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{Generate toy panel data}
\usage{
generate_data(obs_N, obs_T, p, mu, RHO)
}
\arguments{
\item{obs_N}{Number of cross-section units.}

\item{obs_T}{Number of time periods.}

\item{p}{Number of regressors/covariates.}

\item{mu}{Vector for individual loadings.}

\item{RHO}{Pairwise correlation coefficient of the regressors.}
}
\value{
List containing the balanced panel data for estimation
 (data_estimation) and inference (data_inference). data_estimation$y contains
 the dependent variables and data_estimation$x the regressors. data_inference
 is in the same spirit with the minor difference that data_inference$y has three
 columns (one for each signal c** = 0, 0.1, 0.2). In both cases, the panels are sorted
 such  that first T observations are those for unit 1, followed by the T
 observations for unit 2, etc.
}
\description{
\code{generate_data} generates data as in simulation section of Linton, O.,
 Ruecker, M., Vogt, M.,  Walsh, C. (2024) "Estimation and Inference in
 High-Dimensional Panel Data Models with Interactive Fixed Effects".
 The example provides the code used to create the data sets
 \bold{data_estimation.rda} and \bold{data_inference.rda} of the package.
}
\examples{
# Simulate an example data set as in Scenario B of the paper with
# N = 50, T = 50, RHO = 0.25, p = 901
#-----------------------------------------------------------------------------

# Set the parameters
#-----------------------------------------------------------------------------
\dontrun{
# Set the number of cross-sections, time periods
obs_N  <- 50
obs_T  <- 50
p <- 901
# Specify the pairwise correlation among the regressors
RHO <- 0.25
# Specify the mean for the factor loadings
mu <- c(1, 1, 1, rep(1, (p-1)))

# Simulate the data
set.seed(2024)
data <- generate_data(obs_N = obs_N, obs_T = obs_T, p,  mu = mu,
                             RHO = RHO)
data_estimation <- data$data_estimation
data_inference <- data$data_inference
save(data_estimation, file = "data_estimation.rda", compress = TRUE)
save(data_inference, file = "data_inference.rda", compress = TRUE)
}
}
