% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdcce_estimator.r
\name{hdcce_estimator}
\alias{hdcce_estimator}
\title{Estimate HD panels with IFE}
\usage{
hdcce_estimator(
  data,
  obs_N,
  obs_T,
  TRUNC = 0.01,
  NFACTORS = NULL,
  variant = "Lasso",
  lambda = NULL,
  NFOLDS = 10,
  foldid = NULL,
  scree_plot = FALSE,
  standardize = TRUE
)
}
\arguments{
\item{data}{List containing the balanced panel data with data$y containing
the dependent variables and data$x the regressors. Both are sorted such
that first T observations are those for unit 1, followed by the T
observations for unit 2, etc.}

\item{obs_N, obs_T}{The number of cross-section units (obs_N) and time series
length (obs_T) in the data.}

\item{TRUNC}{The truncation parameter tau used to estimate the number of
factors. Default is 0.01.}

\item{NFACTORS}{Allows to set the number of factors used in estimation.
Default is NULL so that the data driven choice with TRUNC = 0.01 is used.}

\item{variant}{Choose whether to use the lasso based estimator
(variant = "Lasso" (Default)) or the least squares variant (variant = "LS").}

\item{lambda}{User specified lambda grid.}

\item{NFOLDS}{The number of folds (partitioned along cross-section n) used for
cross-validation (as described in the paper). Default is NFOLDS = 10.
Fold size can vary by one if obs_N is not divisible by NFOLDS.}

\item{foldid}{Integer vector (obs_N * obs_T - dimensional) containing a label
for each sample to determine its fold for CV.}

\item{scree_plot}{Logical variable to indicate whether a scree plot of the
eigendecomposition of \eqn{\Sigma} should be shown. The default is FALSE.}

\item{standardize}{Logical variable to indicate whether glmnet is called
with the standardized projected data. The default is TRUE.}
}
\value{
The function returns the estimated coefficients, the
number of estimated factors and the lasso penalty parameter. More
specifically:
\itemize{
\item If a lambda sequence was supplied or variant = "LS", then the function
returns a list with
\describe{
  \item{$coefs}{The coefficient estimates.}
  \item{$K_hat}{The estimated number of factors. If NFACTORS was
  set by user then this will be returned.}
  }
\item If no lambda sequence was supplied and variant = "Lasso", then the function
returns a list with
 \describe{
  \item{$coefs}{The estimated coefficients.}
  \item{$K_hat}{The estimated number of factors. If NFACTORS was
  set by user then this will be returned.}
  \item{$Lambda}{The lambda_min calculated by cross-validation.}
  }
}
}
\description{
\code{hdcce_estimator} fits the high-dimensional CCE estimator (HD-CCE-estimator)
proposed in Ruecker, M., Vogt, M., Linton, O., Walsh, C. (2025) "Estimation
and Inference in High-Dimensional Panel Data Models with Interactive
Fixed Effects" \doi{10.3982/QE2308}.
}
\examples{
# Load the data set
data("data_estimation")

# Set the dimensions of the data
obs_N <- 20
obs_T <- 20

# Do the estimation
estimate_model <- hdcce_estimator(data = data_estimation, obs_N = obs_N,
                      obs_T = obs_T, TRUNC = 0.01, NFACTORS = NULL,
                      variant = "Lasso", lambda = NULL, NFOLDS = 10,
                      foldid = NULL, scree_plot = FALSE)
print(estimate_model$coefs[c(1:10)])

}
\references{
Ruecker, M., Vogt, M., Linton, O. and Walsh, C. (2025).
  Estimation and inference in high-dimensional panel data models with
  interactive fixed effects. \emph{Quantitative Economics}, 16(4),
  1457--1509. \doi{10.3982/QE2308}
}
