% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdcce_inference.r
\name{hdcce_inference}
\alias{hdcce_inference}
\title{Inference for HD panels with IFE}
\usage{
hdcce_inference(
  data,
  obs_N,
  obs_T,
  TRUNC = 0.01,
  NFACTORS = NULL,
  NFOLDS = 10,
  foldid = NULL,
  COEF_INDEX_VEC,
  alpha = c(0.01, 0.05, 0.1),
  HAC = 2,
  standardize = TRUE
)
}
\arguments{
\item{data}{List containing the balanced panel data with data$y containing
the dependent variables and data$x the regressors. Both are sorted such
that first T observations are those for unit 1, followed by the T
observations for unit 2, etc.}

\item{obs_N, obs_T}{The number of cross-section units (obs_N) and time series
length (obs_T) in the data.}

\item{TRUNC}{The truncation parameter tau used to estimate the number of
factors. Default is 0.01.}

\item{NFACTORS}{Allows to set the number of factors use in estimation.
Default is NULL so that the data driven choice with TRUNC = 0.01 is used.}

\item{NFOLDS}{Specifies the number of folds to use when using cross-validation
to select the lasso penalty based on cv.glmnet. Default is 10.
Fold size can vary by one if obs_N is not divisible by NFOLDS.}

\item{foldid}{Integer vector (obs_N*obs_T - dimensional) containing a label
for each sample to determine its fold for CV.}

\item{COEF_INDEX_VEC}{Indeces of regressors to be tested.}

\item{alpha}{Vector containing significance levels.}

\item{HAC}{Integer = 1,2,3. 1 employ homoscedastic no serial correlation,
2 heteroscedastic no serial correlation and 3 HAC variance
estimator. 2 is set as default.}

\item{standardize}{Logical variable to indicate whether glmnet is called
with the standardized projected data. The default is TRUE.}
}
\value{
The function returns, for each entry of COEF_INDEX_VEC, a list with
\describe{
  \item{coef_despar}{Desparsified lasso estimate.}
  \item{se}{Estimated standard error of the desparsified estimator.}
  \item{confidence_band}{Symmetric confidence interval (lower and upper
  interval value) for each significance level in alpha.}
}
}
\description{
\code{hdcce_inference} High-dimensional inference with the desparsified HD-CCE
estimator proposed in Ruecker, M., Vogt, M., Linton, O., Walsh, C. (2025)
"Estimation and Inference in High-Dimensional Panel Data Models with Interactive
Fixed Effects" \doi{10.3982/QE2308}.
}
\examples{
# Load the data set
data("data_inference")

# Set the dimensions of the data
obs_N <- 20
obs_T <- 20

# signal = 1,2,3 corresponds to c** = 0, 0.1, 0.2
signal <- 1

# Use correct column of Y
data_inference <- list(x = data_inference$x, y = data_inference$y[,signal])

# Regressor to be tested
COEF_INDEX_VEC <- 1

# Inference example
inference_model <- hdcce_inference(data_inference, obs_N, obs_T, TRUNC = 0.01,
 NFACTORS = NULL, NFOLDS = 10, foldid = NULL, COEF_INDEX_VEC,
 alpha = c(0.01, 0.05, 0.1))

print(inference_model$confidence_band)

}
\references{
Ruecker, M., Vogt, M., Linton, O. and Walsh, C. (2025).
  Estimation and inference in high-dimensional panel data models with
  interactive fixed effects. \emph{Quantitative Economics}, 16(4),
  1457--1509. \doi{10.3982/QE2308}
}
