% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdcce_inference.r
\name{hdcce_inference}
\alias{hdcce_inference}
\title{Inference for HD panels with IFE}
\usage{
hdcce_inference(
  data,
  obs_N,
  obs_T,
  TRUNC = 0.01,
  nFactors = NULL,
  nFolds = NULL,
  foldid = NULL,
  COEF_INDEX,
  alpha = c(0.01, 0.05, 0.1)
)
}
\arguments{
\item{data}{List containing the balanced panel data with data$y containing
the dependent variables and data$x the regressors. Both are sorted such
that first T observations are those for unit 1, followed by the T
observations for unit 2, etc.}

\item{obs_N, obs_T}{The number of cross-section units (obs_N) and (obs_T) in
the data.}

\item{TRUNC}{The truncation parameter tau used to estimate the number of
factors. Default is 0.01.}

\item{foldid}{Integer vector (obs_N*obs_T - dimensional) containing a label
for each sample to determine its fold for CV.}

\item{COEF_INDEX}{Index of regressor to be tested.}

\item{alpha}{Vector containing significance levels.}

\item{NFACTORS}{Allows to set the number of factors use in estimation.
Default is NULL so that the data driven choice with TRUNC=0.01 is used.}

\item{NFOLDS}{Specify the number of folds to use when using cross-validation
to select lasso penalty based on cv.glmnet. Default is the default in
cv.glmnet, which is 10.}
}
\value{
The function returns the following:
\itemize{
 \describe{
  \item{$coef_despar}{Desparsified lasso estimate.}
  \item{$Avar}{Asymptotic variance of desparsified lasso.}
  \item{$conf_bands}{Symmetric confidence bands for given alpha.}
  \item{$var_est_Lasso}{Estimated variance of idiosyncratic regression error.}
  }
}
}
\description{
\code{hdcce_inference} High-dimensional inference with the desparsified CCE
estimator proposed in Linton, 0., Ruecker, M., Vogt, M., Walsh, C. (2024)
"Estimation and Inference in High-Dimensional Panel Data Models with Interactive
Fixed Effects".
}
\examples{
# Load the data set
data("data_example")

# Set the dimensions of the data
obs_N <- 50
obs_T <- 50

# Do the estimation
estimate_model <- ???

}
\references{
Linton, 0., Ruecker, M., Vogt, M., Walsh, C. (2024) "Estimation
and Inference in High-Dimensional Panel Data Models with Interactive
Fixed Effects"
}
